+++ sdk/cosmos/azure_data_cosmos/src/clients/database_client.rs
@@ -70,6 +70,7 @@ impl DatabaseClient {
     ///     .into_model()?;
     /// # }
     /// ```
+    #[tracing::instrument(skip_all, fields(id = self.database_id))]
     pub async fn read(
         &self,
         options: Option<ReadDatabaseOptions<'_>>,
@@ -105,6 +106,7 @@ impl DatabaseClient {
     /// ```
     ///
     /// See [`Query`] for more information on how to specify a query.
+    #[tracing::instrument(skip_all, fields(id = self.database_id))]
     pub fn query_containers(
         &self,
         query: impl Into<Query>,
@@ -129,6 +131,7 @@ impl DatabaseClient {
     /// # Arguments
     /// * `properties` - A [`ContainerProperties`] describing the new container.
     /// * `options` - Optional parameters for the request.
+    #[tracing::instrument(skip_all, fields(id = self.database_id))]
     pub async fn create_container(
         &self,
         properties: ContainerProperties,
@@ -152,6 +155,7 @@ impl DatabaseClient {
     ///
     /// # Arguments
     /// * `options` - Optional parameters for the request.
+    #[tracing::instrument(skip_all, fields(id = self.database_id))]
     pub async fn delete(
         &self,
         options: Option<DeleteDatabaseOptions<'_>>,
@@ -170,10 +174,11 @@ impl DatabaseClient {
     ///
     /// # Arguments
     /// * `options` - Optional parameters for the request.
+    #[tracing::instrument(skip_all, fields(id = self.database_id))]
     pub async fn read_throughput(
         &self,
         options: Option<ThroughputOptions<'_>>,
-    ) -> azure_core::Result<Option<Response<ThroughputProperties>>> {
+    ) -> azure_core::Result<Option<ThroughputProperties>> {
         let options = options.unwrap_or_default();
 
         // We need to get the RID for the database.
@@ -193,6 +198,7 @@ impl DatabaseClient {
     /// # Arguments
     /// * `throughput` - The new throughput properties to set.
     /// * `options` - Optional parameters for the request.
+    #[tracing::instrument(skip_all, fields(id = self.database_id))]
     pub async fn replace_throughput(
         &self,
         throughput: ThroughputProperties,