+++ sdk/storage/azure_storage_blob/tests/page_blob_client.rs
@@ -212,6 +212,7 @@ async fn test_set_sequence_number(ctx: TestContext) -> Result<(), Box<dyn Error>
 }
 
 #[recorded::test]
+#[ignore = "https://github.com/Azure/azure-sdk-for-rust/issues/3441"]
 async fn test_upload_page_from_url(ctx: TestContext) -> Result<(), Box<dyn Error>> {
     // Recording Setup
     let recording = ctx.recording();