+++ sdk/core/typespec_client_core/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added `Request::body_mut()`.
 - Added `UrlExt::set_query_pair()` to simplify overwriting query parameter key values.
+- Sort query parameters lexicographically in `QueryBuilder`.
 
 ### Breaking Changes